refactor: Move node logic to node.rs and remove unused
