+++ src/blockchain/error.rs
@@ -1,5 +1,7 @@
 use bitcoin::{blockdata::script, BlockHash};
 use btcd_rpc::error::UtreexodError;
+
+use super::p2p_blockchain::node::NodeRequest;
 #[derive(Debug)]
 pub enum BlockchainError {
     BlockNotPresent,
@@ -17,7 +19,14 @@ pub enum BlockchainError {
     MessageTooBig,
     #[cfg(feature = "experimental-p2p")]
     PeerMessageInvalidMagic,
+    #[cfg(feature = "experimental-p2p")]
+    RequestTimeout,
+    #[cfg(feature = "experimental-p2p")]
+    NoPeersAvailable,
+    #[cfg(feature = "experimental-p2p")]
+    ChannelError(async_std::channel::SendError<NodeRequest>),
 }
+
 #[derive(Debug)]
 pub enum BlockValidationErrors {
     PrevBlockNotFound(BlockHash),
@@ -34,6 +43,11 @@ impl From<bitcoin::consensus::encode::Error> for BlockchainError {
 impl From<kv::Error> for BlockchainError {
     fn from(err: kv::Error) -> Self {
         BlockchainError::DatabaseError(err)
+    }
+}
+impl From<async_std::channel::SendError<NodeRequest>> for BlockchainError {
+    fn from(err: async_std::channel::SendError<NodeRequest>) -> Self {
+        BlockchainError::ChannelError(err)
     }
 }
 impl From<UtreexodError> for BlockchainError {
