+++ src/blockchain/p2p_blockchain/mempool.rs
@@ -1,0 +1,4 @@
+pub struct Mempool;
+impl Mempool {
+    pub fn accept_to_mempool(&self) {}
+}
